+++ retro/import/__main__.py
@@ -1,4 +1,0 @@
-from retro.scripts.import_path import main
-
-if __name__ == "__main__":
-    main()