+++ scripts/import.py
@@ -1,4 +1,4 @@
 #!/usr/bin/env python
-from retro.scripts.import_path import main
+from stable_retro.scripts.import_path import main
 
 main()