+++ scripts/import_sega_classics.py
@@ -1,4 +1,4 @@
 #!/usr/bin/env python
-from retro.scripts.import_sega_classics import main
+from stable_retro.scripts.import_sega_classics import main
 
 main()