+++ stable_retro/import/__main__.py
@@ -1,0 +1,4 @@
+from stable_retro.scripts.import_path import main
+
+if __name__ == "__main__":
+    main()